Socket Architecture improved to follow current MVC pattern
